src/main/requirements.txt
click==8.0.1
Flask==2.0.1
PyYaml==5.4.1
flask-cors==3.0.10
python-dateutil==2.8.2
tornado==6.1